src/requirements.txt
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.2
numpy==1.26.4
pymongo==4.6.1
schedule==1.2.1
Flask==3.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0 
flask-caching==1.10.1